Making Documentation Available from the Service
Rocket_okapi was added as a dependency for the netspot_control project.
It is used to generate OpenAPI documentation from the implementation and
to present both design and generated OpenAPI documentation with RapiDoc
and Swagger UI.

The main rust file was updated to use rocket_okapi. In addition,
network-related endpoint implementation was moved to the network.rs
file.

Index.html for server root is now in the static folder. This folder also
contains favicon and designed OpenAPI.

Dockerfile was updated to include static files for the netspot_control
server.

The roadmap was renamed to TODO and updated for README.