Adding TOML Generator for Netspot
This commit is mainly for adding TOML generation from the NetspotConfig,
which allows us to write configuration files for the netspot processes.

Netspot configuration AVG_PKG_SIZE fixed to the correct name of
AVG_PKT_SIZE.

TODO listing was updated for the README.

OpenAPI HTML document was removed as the documentation is available from
the service itself when it is run.