Refactoring and JsonSchemas
This commit is updating the netspot_control module structure as
preparation for future implementations. Additionally, JsonSchema derive
was added for NetspotConfig and its sub-structures. These JsonSchemas
allow rocket_okapi to generate schemas when structures are used in
endpoints.

The default netspot configuration JSON constant was made available from
the netspot module as it will be used when writing the default
configuration into the database when the database is created on the
first run.