State Manager with Database
This commit adds NetspotControl structure, which works as the heart of
the service and is planned to provide access to different application
parts for HTTP API requests. In addition, the initial Database solution
is implemented to make configuration requests possible.

This commit adds the Diesel crate for database handling and migration.
In addition, migration scripts were added to build the initial database
with the default configuration.

HTTP API was updated to include an endpoint for getting configuration
from the database.

The dotenv crate was added for configuring the database path in the
'.env' file for development. Dockerfile was updated to include SQLite
library in the final image and database path environment variable was
set to "/var/lib/netspot_control/netspot_control.db".

The program component diagram was updated slightly.