Configuration Endpoints Complete
This commit adds endpoints for posting new configurations, updating
existing configurations, and deleting configurations with id.

The Database implementation was updated to allow the actions mentioned
above. In addition, a get_configurations function was added, which is
needed to read all configurations to NetspotManager when the application
starts.

The OpenAPI design was updated to version 0.1.1 as the return value from
the configuration POST was removed.