Netspot Statuses Endpoint
This commit adds an endpoint to get a list of available netspot
configurations and their run status. Netspot process is not fully
implemented, and the running status is faked by checking if the netspot
configuration TOML file exists. The TOML file is created from the
NetspotProcess start method and removed from the stop method.

Configuration endpoints were improved to update netspot configurations
for the NetspotManager. Current solutions updates all configurations
which in many cases are not necessary. Better solutions should be added
later.

Refactoring was done by renaming the configurations module as structures
and adding new structures modules there.