Moving Project Files to the Root of Repository
These changes are preparations for the CI scripts we will generate with
the 'sifis-generate' tool. All project files from the netspot_control
folder are moved to the root of this repository.

The Dockerfile was updated to work with the changes, and README was
updated to tell more about project folders. In addition, a new section
was added to instruct how the database can be made persistent with
Docker.