Status Endpoints Complete
This commit adds all of the rest of the planned endpoints for handling
netspot statuses. However, the NetspotManager and NetspotProcess are not
yet done; therefore, the running status is fake information based on the
availability of the netspot process configuration TOML file.

Updated the TODO list in the README and removed some old entries.