Removing and Fixing Actions
This commit removes the deploy script as we do not deploy the
application yet. Furthermore, the wp4-edge_ids.yml was deleted as the
same jobs are also run from the Ubuntu script. Finally, the Ubuntu
script was fixed to install libpcap for all the jobs, and project names
were renamed to netspot_control.