Continuous integration (#2)
* Continuous Integration

This commit adds files generated with the 'sifis-generate' tool.

* GitLab File Not Needed

* Removing Windows and macOS Deployment Scripts

This commit removed Windows and macOS deployment scripts, which were not
tested to work. Additionally, installation for the needed libpcap
development library was added for the Ubuntu deployment script.

* Removing and Fixing Actions

This commit removes the deploy script as we do not deploy the
application yet. Furthermore, the wp4-edge_ids.yml was deleted as the
same jobs are also run from the Ubuntu script. Finally, the Ubuntu
script was fixed to install libpcap for all the jobs, and project names
were renamed to netspot_control.

* Running apt with the sudo

Sudo was added for all software installation steps in the hope of fixing
the permission denied error.

* Adding Action to Build Docker Image

This script could be later changed to upload releases.

* Disabling Code Coverage Until More Tests Are Added

We are temporarily removing the code coverage job so that we can benefit
from other scripts created by the sifis-generate tool for code review.

Code coverage should be enabled after we have time to add more tests.

* Re-enabling Some of the Code Coverage

Most of the code coverage is re-enabled as recommended. Additionally, we
renamed netspot_control-ubuntu to netspot_control, as that name makes

README updated to use the correct action badge.