Listening and Establishing Netspot Connections
With these changes, netspot can already be connected to the service
using a generated configuration file. Service prints received messages
to stdout for now but are planned to be forwarded to the database writer
via a suitable channel and later to the webhook handler.

We did not find an ideal way to handle tasks from the previous Fairing
solution and, therefore, have returned to the managed state. Tokio
channels are used to broadcast a shutdown message for all worker tasks
and to wait for them to complete their work.