Starting and Stopping Netspot
This commit adds functionality to start and stop netspot processes on
demand. Netspot shutdown is done using a SIGINT signal. If the netspot
process, however, does not shut down correctly in 5 seconds, it is shut
forcefully.

The delete configuration endpoint was updated to stop the netspot
process before deleting its configuration.

Two new packages were added. The Tokio was already partly included
through Rocket, but it is now directly included as we need a process
feature from it. In addition, the Nix crate was added for sending SIGINT
signals for netspot processes to shut them down gracefully when
possible.