Forwarding Stats from Netspot
This commit adds statistics structures and a broadcast channel for
forwarding messages from netspot processes to the message handlers. In
addition, a test handler was added for debug builds that prints messages
to the standard output.

Netspot handler tasks were updated to use the above broadcast channel to
forward correctly received messages.

The application diagram was updated as the new plan is to simplify the
application by writing single-worker tasks for writing netspot messages
to the database or sending them with webhooks.

The Termion package was added to allow colors for terminal output.

Commit also includes some preliminary work to add a message handler to
write messages to the database.