Adding Database Writing
This commit adds implementation to write netspot results to the
database. The implementation keeps one hour of data in the cache
database and removes older measurements periodically.

OpenAPI specification was updated slightly as statistic endpoints are
planned to use time as milliseconds since epoch instead of nanosecond
accuracy. However, the nanosecond accuracy is still used in the results.
In addition, the messages time type was changed to i128 to ensure we do
not run out of bits.

The timestamp column was changed from INTEGER to BIGINT on message
tables on the database. This change was needed, so the diesel uses the
i64 type instead of the i32.

NetspotProcess was fixed to return the "Running" status when the process
has been marked disabled but has not been stopped.