Selectable Netspot Output
Service now checks if the SHOW_NETSPOT_MESSAGES environment variable is
set and is a non-zero value. If that is true, netspot messages are
printed to the stdout. This environment variable is set from the '.env'
file to show messages when developing the service.

Loading the '.env' file is now done at the start of the main function.