Adding Example Webhook Server
This commit adds a simple HTTP server, which prints out received
messages to the standard output. This server is planned to be used to
test netspot control webhooks, but it is also an example of how to
create an application able to receive netspot messages.

The Python folder in the example folder was split into two. One folder
is for the example client application, and another is for the example
webhook server.

Webhook structure serialization was improved to skip the headers field
if it is empty.