Adding Webhook Sender
This commit adds WebhookManager with worker tasks for sending netspot
messages to registered webhooks.

The reqwest package was added for sending netspot messages.

OpenAPI specification was modified to match implementation, and the
version number was changed to 1.0.0 as the service now has all the
planned features for the first release.

The Python webhook example server was updated to silence httpd log
messages to only print received netspot messages. In addition, header
printing was used to check that netspot_control service sents configured
extra headers correctly. This code line for header printing is included
but commented out.

Webhook endpoints were improved to send the updated configuration to the
webhooks components when necessary.

The database component was updated with a function to read all webhook
configurations from the cache database.