Influxdb (#3)
* Adding InfluxDB Structure

This commit adds InfluxDB1Config with unit tests. Default values are the
same that the netspot program uses.

The version was changed to 1.0.2.

* InfluxDB Configuration TOML

This commit inserts the InfluxDB1Config configuration as an optional
part of the NetspotConfig. In addition, the configuration is added to
netspot TOML and JSON when it is defined.

Unit tests were updated to check new feature is functioning correctly.

* Fixing InfluxDB Exporter

This commit removes '1' from the exporter name. The correct name is
without the one.

* Fixing Unit Tests and Updating

The configuration unit test was fixed expect the correct InfluxDB
exporter.

Dockerfile was updated to install upgrades for build images and the
final image.

Some of the containers were updated to newer versions.

* Clippy Fixes