Async NetspotManager
This commit changes almost all NetspotManager functions to async. In
addition, the RwLock for the NetspotManger had to be changed from
std::sync to tokio::sync, which is more expensive but can be held across
an await. These changes now ensure that the netspot process is stopped
correctly when the server is shut down.

Database cleanup messages were also made consistent with other messages.