Testing Configuration Endpoints
Configuration endpoints are now tested for CRUD and invalid requests.

The NetspotProcess was updated to kill the netspot process on the drop,
which is essential for cases if any of the HTTP API unit tests fail.
Otherwise, the netspot process would keep running because the cleanup
routine is not run as objects are only dropped.

Test setup functions were moved as implementations for the TestSetup
struct.