Testing Server Binary
The server was improved to allow some command-line arguments. These
arguments allow us to run server tests in a temporary directory.

A new integration test was added to test the server binary in its normal
running mode. It is testing for a successful launch, waiting for data
message, sending an alarm and waiting for the alarm message, and
finally, for graceful shutdown. Webhook testing is still to be done.

In addition, this commit changed Cargo.toml to use crates based on their
major and minor version. The server software version was updated to
1.1.0 as we slightly updated the API (to meet the planned design).