Known Issues
This commit adds known problems to the README file that cannot be fixed
now. The problematic tests are bypassed with address sanitizer and
Valgrind. It is hoped that these can be fixed, but other crucial tasks
in another work package must be done first.

The server binary now allows automatic shutdown after a given amount of
seconds. This feature is planned for testing the server with Valgrind.