Testing (#5)
* Testing Preparation

This commit contains preliminary code for testing endpoints. While doing
this, problems with threads and task management were encountered, and
therefore the commit also includes fixes for them. The netspot process
management still needs more work, but we want to save the progress
before that.

* Async NetspotManager

This commit changes almost all NetspotManager functions to async. In
addition, the RwLock for the NetspotManger had to be changed from
std::sync to tokio::sync, which is more expensive but can be held across
an await. These changes now ensure that the netspot process is stopped
correctly when the server is shut down.

Database cleanup messages were also made consistent with other messages.

* Customizable Data Path

For unit testing, we need to be able to output all files to a temporary
directory. These changes make that possible.

* Testing Configuration Endpoints

Configuration endpoints are now tested for CRUD and invalid requests.

The NetspotProcess was updated to kill the netspot process on the drop,
which is essential for cases if any of the HTTP API unit tests fail.
Otherwise, the netspot process would keep running because the cleanup
routine is not run as objects are only dropped.

Test setup functions were moved as implementations for the TestSetup
struct.

* Testing Network and Statistics Endpoints

This commit adds unit testing for Network and Statistics endpoints. In
addition, it includes some typo fixes.

* Testing Status Endpoints

This commit adds testing for all status endpoints with valid and invalid
requests.

A new function was added to the tests_common module for converting the
Statuses vector to a handy HashMap. This function is essential since
there is no guarantee for returned statuses to be in numeric order.

* Testing the Testing Endpoint

The testing endpoint allows a developer to send test alarm messages. The
added test sends default and custom alarm messages and checks that they
are inserted into the database.

* Testing Webhook Endpoints

This commit adds unit testing for Webhook endpoints with valid and
invalid requests.

IMPORTANT: Webhook listing endpoint URI was changed from
"/netspot/webhooks" to "/netspots/webhooks" to match the naming
convention of the other endpoints of the webhook group. It was also
originally the planned name for the endpoint.

* Preparing for More Testing

The tests_common module was moved under main. In addition, the test
folder was renamed as tests for rust standard integral tests to be
added. Finally, compiled python files were removed.

* Testing Server Binary

The server was improved to allow some command-line arguments. These
arguments allow us to run server tests in a temporary directory.

A new integration test was added to test the server binary in its normal
running mode. It is testing for a successful launch, waiting for data
message, sending an alarm and waiting for the alarm message, and
finally, for graceful shutdown. Webhook testing is still to be done.

In addition, this commit changed Cargo.toml to use crates based on their
major and minor version. The server software version was updated to
1.1.0 as we slightly updated the API (to meet the planned design).

* Webhook Testing

This commit contains implementation for testing webhooks.

Actix Web crate was added for testing as the server test uses it to
receive netspot messages with webhooks.

* Github Action Update

This commit adds netspot installation for code coverage tests as needed
for unit and integration tests. In addition, the evaluation of code
coverage value is now enabled.

* Running Install with Sudo

* Setting Capabilities for Netspot

This commit updates GitHub action to set capabilities for netspot binary
so that when it is run in testing, there are enough permissions to
capture network traffic, thus allowing the integration test to pass.

* Installing Netspot for More Tests

This commit adds install for netspot binary for careful-ubuntu and
address-sanitizer-ubuntu.

* Known Issues

This commit adds known problems to the README file that cannot be fixed
now. The problematic tests are bypassed with address sanitizer and
Valgrind. It is hoped that these can be fixed, but other crucial tasks
in another work package must be done first.

The server binary now allows automatic shutdown after a given amount of
seconds. This feature is planned for testing the server with Valgrind.